# Avoides allocation with constExpr execution