Previously executConstExpr is called various places, and therefore it was necessary to return `interface{}`
and that resulted in allocation per call.
This commit avoids the allocation by adding `executConstExprI32` which is used by data segment manipulation,
and repurposes `executConstExpr` solely to global instance initialization.

As a result, this completely removes the allocation around const expr execution, and hence the perf improvement
in the instantiation phrase. The improvement diff is proportionate to the number of data segments and globals.

```
$ benchstat old.txt new.txt
name                                    old time/op    new time/op    delta
Initialization/interpreter-32             36.0µs ± 1%    36.1µs ± 3%    ~     (p=0.481 n=26+26)
Initialization/interpreter-multiple-32    38.5µs ± 9%    37.9µs ±10%    ~     (p=0.114 n=29+30)
Initialization/compiler-32                28.7µs ± 9%    27.9µs ±10%  -2.94%  (p=0.019 n=30+30)
Initialization/compiler-multiple-32       25.1µs ± 9%    23.9µs ± 6%  -4.69%  (p=0.000 n=30+29)
Compilation/with_extern_cache-32           197µs ± 1%     197µs ± 1%    ~     (p=0.971 n=30+30)
Compilation/without_extern_cache-32       5.89ms ± 1%    5.91ms ± 1%  +0.30%  (p=0.021 n=30+30)

name                                    old alloc/op   new alloc/op   delta
Initialization/interpreter-32              136kB ± 0%     136kB ± 0%  -0.01%  (p=0.000 n=24+30)
Initialization/interpreter-multiple-32     137kB ± 0%     137kB ± 0%  -0.01%  (p=0.000 n=27+27)
Initialization/compiler-32                 137kB ± 0%     137kB ± 0%  -0.01%  (p=0.000 n=26+23)
Initialization/compiler-multiple-32        142kB ± 0%     142kB ± 0%  -0.01%  (p=0.000 n=26+27)
Compilation/with_extern_cache-32          54.6kB ± 0%    54.6kB ± 0%  +0.01%  (p=0.022 n=30+30)
Compilation/without_extern_cache-32       1.99MB ± 0%    1.99MB ± 0%    ~     (p=0.416 n=26+30)

name                                    old allocs/op  new allocs/op  delta
Initialization/interpreter-32               38.0 ± 0%      35.0 ± 0%  -7.89%  (p=0.000 n=30+30)
Initialization/interpreter-multiple-32      57.0 ± 0%      54.0 ± 0%  -5.26%  (p=0.000 n=30+30)
Initialization/compiler-32                  38.0 ± 0%      35.0 ± 0%  -7.89%  (p=0.000 n=30+30)
Initialization/compiler-multiple-32         47.0 ± 0%      44.0 ± 0%  -6.38%  (p=0.000 n=30+30)
Compilation/with_extern_cache-32             982 ± 0%       982 ± 0%    ~     (all equal)
Compilation/without_extern_cache-32        32.6k ± 0%     32.6k ± 0%    ~     (p=0.913 n=30+30)
```