@mathetake 2023-03-20
> Compilation/without_extern_cache-32       5.89ms ± 1%    5.91ms ± 1%  +0.30%  (p=0.021 n=30+30)

this must be the noise - this commit doesn't change any code path of compilation.